Fix SICP context abstraction
